changed hostID to uint8 outside of IPAM

+++ plugins/contiv/ipam.go
@@ -32,7 +32,7 @@ type IPAM struct {
 	sync.RWMutex
 
 	//TODO apply podSubnetCIDR to IPAM structures properly
-	podSubnetCIDR       string // subnet from which individual pod networks are allocated
+	podSubnetCIDR string // subnet from which individual pod networks are allocated
 
 	hostID                uint8           // identifier of host node for which this IPAM is created for
 	podNetworkIPPrefix    net.IPNet       // IPv4 subnet prefix for all pods of one host node (given by hostID)
@@ -47,13 +47,12 @@ type IPAMConfig struct {
 	PodNetworkPrefixLen uint8  // prefix length of subnet used for all pods of 1 host node (pod network = pod subnet for one 1 host node)
 }
 
-//TODO fix hostID from uint8 to uint32
 // newIPAM returns new IPAM module to be used on the host.
-func newIPAM(logger logging.Logger, hostID uint32, config *IPAMConfig) (*IPAM, error) {
+func newIPAM(logger logging.Logger, hostID uint8, config *IPAMConfig) (*IPAM, error) {
 	// create basic IPAM
 	ipam := &IPAM{
-		Logger: logger,
-		hostID: hostID,
+		Logger:        logger,
+		hostID:        hostID,
 		podSubnetCIDR: config.PodSubnetCIDR,
 	}
